+++ cynic-introspection/src/lib.rs
@@ -1,3 +1,4 @@
-mod query;
+pub mod query;
+mod schema;
 
-pub use query::*;
+pub use schema::*;